+++ app/build.gradle
@@ -12,7 +12,7 @@ dependencies {
 }
 
 android {
-    compileSdkVersion 23
+    compileSdkVersion 22
     buildToolsVersion "23.0.1"
 
 
@@ -23,9 +23,9 @@ android {
     defaultConfig {
         applicationId 'ee.ioc.phon.android.speak'
         minSdkVersion 14
-        targetSdkVersion 23
-        versionCode 1313
-        versionName '1.3.13'
+        targetSdkVersion 22
+        versionCode 1315
+        versionName '1.3.15'
     }
